modified configure.sh
Added a few enhancements, in order to allow the script to be run more than once in a correct way. The original script in fact doesn't remove previously set values where no value is entered during a subsequent run: this causes e.g. already set open ports to remain open, even if no value is provided by the user. Besides, the user is not aware of that. I believe that this might entail some security issues and that's why I edited the functions 'get_conf_var' and 'change_conf_var', as well as the test for 'INT_IF'. Now when a value is _not_ entered to the script, and no defaults are set in the script itself either, then previously set values are removed, unless they are commented.

In addition to the above, the script header was also edited to fit better in a 80 character terminal.