Update configure.sh
Hello Arno

> Please put the read on the next line for clarity

Done.

I tried to squash the commits with git, but I couldn't do that, as I made all changes using GitHub's web interface, so I have nothing locally. In fact my question was if it's ever possible to do the squash with this interface. I do understand that you prefer to use git for the development of aif, but right now I'm not really motivated to learn it. If you need a one-file-patch in order to merge my changes, then I can send it to you with an old-fashioned e-mail; if you're still interested in them of course. :)